# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ClinicFlow API Documentation",
        "description": "Comprehensive API documentation for ClinicFlow Backend",
        "contact": {
            "email": "admin@clinicflow.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        },
        {
            "url": "https://api.clinicflow.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/mobile/doctor/appointments": {
            "get": {
                "tags": [
                    "Mobile - Doctor"
                ],
                "summary": "Get doctor's appointments",
                "operationId": "a89b7fc4875a405ba7779da0b9e9d951",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointments list"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mobile/doctor/appointments/today": {
            "get": {
                "tags": [
                    "Mobile - Doctor"
                ],
                "summary": "Get today's appointments",
                "operationId": "641ada6ad20950a65094d6b5ea5757c5",
                "responses": {
                    "200": {
                        "description": "Today's appointments"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mobile/doctor/appointments/{id}/status": {
            "put": {
                "tags": [
                    "Mobile - Doctor"
                ],
                "summary": "Update appointment status",
                "operationId": "9c069bf97bfddf6104b4982f14411af7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "confirmed",
                                            "completed",
                                            "cancelled"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mobile/doctor/patients": {
            "get": {
                "tags": [
                    "Mobile - Doctor"
                ],
                "summary": "Get doctor's patients",
                "operationId": "94d89648b4c31a32f319557f3f939351",
                "responses": {
                    "200": {
                        "description": "Patients list"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mobile/doctor/schedule": {
            "get": {
                "tags": [
                    "Mobile - Doctor"
                ],
                "summary": "Get doctor schedule",
                "operationId": "cb615f3d072159387b5d9572a42a3c49",
                "responses": {
                    "200": {
                        "description": "Doctor schedule"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Mobile - Doctor"
                ],
                "summary": "Update doctor schedule",
                "operationId": "e6b2b884ac9e8ad11f0d71c9ff0a647e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "available_from",
                                    "available_to"
                                ],
                                "properties": {
                                    "available_from": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "09:00"
                                    },
                                    "available_to": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "17:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Schedule updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mobile/doctor/statistics": {
            "get": {
                "tags": [
                    "Mobile - Doctor"
                ],
                "summary": "Get doctor statistics",
                "operationId": "d3150de3db8ab4a468d852d6a97aa8b8",
                "responses": {
                    "200": {
                        "description": "Doctor statistics"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}